+++ src/collectionset.js
@@ -7,168 +7,22 @@ var _        = require( 'underscore' ),
     onAll    = Events.onAll,
     offAll    = Events.offAll;
 
-function sortedIndex( array, obj, iteratee, context ){
-    if( typeof iteratee === 'function' && iteratee.length == 2 ){
-        var value = obj;
-        var low = 0, high = array.length;
-        while (low < high) {
-            var mid = Math.floor((low + high) / 2);
-            if (iteratee.call( context, array[mid], value) < 0 ) low = mid + 1; else high = mid;
-        }
-        return low;
-    }
-    else return _.sortedIndex( array, obj, iteratee, context );
-}
+
 
 module.exports = {
     fastCopy    : fastCopy,
     toModel     : toModel,
     addOne      : addOne,
+    addMany     : addMany,
     removeOne   : removeOne,
     removeMany  : removeMany,
     setMany     : setMany,
-    replaceMany : replaceMany
+    emptySetMany : replaceMany
 };
 
-// Copy options as fast as its possible.
-function fastCopy( dest, source ){
-    if( source ){
-        for( var i in source ){
-            dest[ i ] = source[ i ];
-        }
-    }
 
-    return dest;
-}
 
-// convert argument to model. Return false if fails.
-function toModel( collection, attrs, a_options ){
-    // Only subtype of current collection model is allowed
-    var Model = collection.model;
-    if( attrs instanceof Model ) return attrs;
 
-    var options        = fastCopy( {}, a_options );
-    options.collection = collection;
-    var model          = new Model( attrs, options );
-
-    if( model.validationError ){
-        trigger3( collection, 'invalid', collection, model.validationError, options );
-        return false;
-    }
-
-    return model;
-}
-
-// fast-path for singular add and remove...
-function addOne( collection, el, options ){
-    var model = collection.get( el );
-    if( model ){
-        return model;
-    }
-
-    model = toModel( collection, el, options );
-    if( model ){
-        var models = collection.models,
-            at     = options.at;
-
-        if (at != null){
-            at = +at;
-            if (at < 0) at += this.length + 1;
-            if( at < 0 ) at = 0;
-            if( at > this.length ) at = this.length;
-        }
-
-        if( collection.comparator && at == null && options.sort !== false ){
-            at = sortedIndex( models, model, collection.comparator, collection );
-        }
-
-        if( at ){
-            models.splice( at, 0, model );
-        }
-        else{
-            models.push( model );
-        }
-
-        _addIndex( collection._byId, model );
-        _addReference( collection, model );
-
-        if( !options.silent ){
-            trigger3( model, 'add', model, collection, options );
-            trigger2( collection, 'update', collection, options );
-        }
-
-        return model;
-    }
-}
-
-function removeOne( collection, el, options ){
-    var model = collection.get( el );
-    if( model ){
-        var models = collection.models,
-            sorted = collection.comparator && options.sort !== false;
-
-        var at = sorted ?
-                 sortedIndex( models, model, collection.comparator, collection )
-            : _.indexOf( models, model );
-
-        models.splice( at, 1 );
-
-        _removeIndex( collection._byId, model );
-
-        if( !options.silent ){
-            trigger3( model, 'remove', model, collection, options );
-            trigger2( collection, 'update', collection, options );
-        }
-
-        _removeReference( collection, model );
-
-        return model;
-    }
-}
-
-/** Optimized for removing many elements
- * 1. Remove elements from the index, checking for duplicates
- * 2. Create new models array matching index
- * 3. Send notifications and remove references
- */
-function removeMany( collection, toRemove, a_options ){
-    var _byId = collection._byId;
-
-    // 1. Remove models from index
-    var removed = Array( toRemove.length ),
-        i, j;
-
-    for( i = 0, j = 0; i < toRemove.length; i++ ){
-        var model = collection.get( toRemove[ i ] );
-        if( model ){
-            removed[ j++ ] = model;
-            _removeIndex( _byId, model );
-        }
-    }
-
-    removed.length = j;
-
-    // 2. Remove models from array
-    var prev   = collection.models,
-        models = collection.models = Array( prev.length - j );
-
-    for( i = 0, j = 0; i < prev.length; i++ ){
-        model = prev[ i ];
-
-        if( _byId[ model.cid ] ){
-            models[ j++ ] = model;
-        }
-    }
-
-    models.length = j;
-
-    // 3. Send notifications and dereference models
-    _removeModels( collection, removed, a_options );
-
-    a_options.silent || !removed.length || trigger2( collection, 'update', collection, a_options );
-
-    return removed;
-}
 
 var silence = { silent : true };
 
@@ -196,33 +50,7 @@ function replaceMany( self, models, a_options ){
     return added;
 }
 
-function emptySetMany( self, models, a_options ){
-    var options = fastCopy( {}, a_options ),
-        notify  = !options.silent;
 
-    _reallocate( self, models, function( source ){
-        var model = toModel( self, source, options );
-
-        if( model ){
-            return _addReference( self, model );
-        }
-    });
-
-    var added = this.models;
-
-    var sort = self.comparator && added.length && options.sort !== false;
-    if( sort ) self.sort( silence );
-
-    if( notify ){
-        added.length && _notifyAdd( self, added, options );
-        sort && trigger2( self, 'sort', self, options );
-        if( added.length || removed.length ){
-            trigger2( self, 'update', self, options );
-        }
-    }
-
-    return added;
-}
 
 function _removeModels( collection, removed, options ){
     var silent = options.silent;
@@ -412,89 +240,6 @@ function fpMerge( collection, keepCount, toKeep, toAdd ){
     return toRemove;
 }
 
-// Update a collection by `set`-ing a new list of models, adding new ones,
-// removing models that are no longer present, and merging models that
-// already exist in the collection, as necessary. Similar to **Model#set**,
-// the core operation for updating the data contained by the collection.
-function setMany2( self, a_models, a_options ){
-    if( a_models == null ) return;
-
-    var options = fastCopy( { merge : true }, a_options ),
-        models  = a_models;
-
-    var merge  = options.merge;
-
-    var sort     = false,
-        sortable = self.comparator && at == null && options.sort !== false,
-        sortAttr = typeof self.comparator == 'string' ? self.comparator : null;
-
-    // Turn bare objects into model references, and prevent invalid models
-    // from being added.
-    var previous = self.models,
-        toAdd = [];
-
-    _reallocate( self, models, function( source, _byCid ){
-        // If a duplicate is found, prevent it from being added and
-        // optionally merge it into the existing model.
-        var existing = self.get( source );
-        if( existing ){
-            if( merge && source !== existing ){
-                var attrs = source.attributes || source;
-                if( options.parse ) attrs = existing.parse( attrs, options );
-                existing.set( attrs, options );
-                if( sortable && !sort ) sort = existing.hasChanged( sortAttr );
-            }
-
-            if( !_byCid[ existing.cid ] ){
-                return existing;
-            }
-        }
-        else{
-            var model = toModel( self, model, options );
-            if( model ){
-                toAdd.push( model );
-                return _addReference( self, model );
-            }
-        }
-    });
-
-    if( sort || ( sortable && toAdd.length ) ){
-        self.sort( { silent : true } );
-    }
-
-    // remove references and fire 'remove' events if needed...
-    if( this.models.length - toAdd.length < previous.length ){
-        _garbageCollect( self, previous, options );
-    }
-
-    // Unless silenced, it's time to fire all appropriate add/sort events.
-    if( !options.silent ){
-        _notifyAdd( self, toAdd, options );
-        if( sort ) trigger2( self, 'sort', self, options );
-        if( toAdd.length || toRemove.length ) trigger2( self, 'update', self, options );
-    }
-
-    // Return the added (or merged) model (or models).
-    return self.models;
-}
-
-// Remove references from models missing in collection's index
-// Send 'remove' events if no silent
-function _garbageCollect( collection, previous, options ){
-    var _byId = collection._byId,
-        silent = options.silent;
-
-    // Filter out removed models and remove them from the index...
-    for( var i = 0; i < previous.length; i++ ){
-        var model = previous[ i ];
-
-        if( !_byId[ model.cid ] ){
-            silent || trigger3( model, 'remove', model, collection, options );
-            _removeReference( collection, model );
-        }
-    }
-}
-
 /**
  * create - factory function to create model from attrs
  * merge  - function to merge attrs and models
@@ -543,84 +288,5 @@ function _mergeModels( self, source, options ){
     return models;
 }
 
-// reallocate model and index
-function _reallocate( self, source, getModel){
-    var models = Array( source.length ),
-        _byId   = {};
-
-    for( var i = 0, j = 0; i < source.length; i++ ){
-        var src = source[ i ];
-        if( src ){
-            var model = getModel( src, _byId );
-            // add to array and indexes...
-            if( model ){
-                models[ j++ ] = model;
-
-                _byId[ model.cid ] = model;
-
-                var id = model.id;
-                if( id ){
-                    _byId[ id ] = model;
-                }
-            }
-        }
-    }
-
-    models.length = j;
-    self.models = models;
-    self._byId = _byId;
-}
-
-/**
- * Helper functions
- */
 
 
-// Ownership and events subscription
-function _addReference( collection, model ){
-    model.collection || ( model.collection = collection );
-    onAll( model, collection._onModelEvent, collection );
-    return model;
-}
-
-function _removeReference( collection, model ){
-    if( collection === model.collection ){
-        model.collection = void 0;
-    }
-    offAll( model, collection._onModelEvent, collection );
-}
-
-// Index management
-function _addIndex( _byId, model ){
-    _byId[ model.cid ] = model;
-    var id             = model.id;
-    if( id != null ){
-        _byId[ id ] = model;
-    }
-}
-
-function _removeIndex( _byId, model ){
-    delete _byId[ model.cid ];
-    var id = model.id;
-    if( id != null ){
-        delete _byId[ id ];
-    }
-}
-
-function _move( source, at, len ){
-    for( var j = source.length - len, i = at; j < source.length; i++, j++ ){
-        var x       = source[ i ];
-        source[ i ] = source[ j ];
-        source[ j ] = x;
-    }
-}
-
-function _notifyAdd( self, models, options ){
-    var at = options.at;
-
-    for( var i = 0; i < models.length; i++ ){
-        var model = models[ i ];
-        if( at != null ) options.index = at + i;
-        trigger3( model, 'add', model, self, options );
-    }
-}
